+++ include/delfem2/femsolidlinear.h
@@ -162,9 +162,9 @@ void MergeLinSys_SolidLinear_Static_MeshTri2D(
     const double g_x,
     const double g_y,
     const double* aXY1,
-    unsigned int nXY,
+    size_t nXY,
     const unsigned int* aTri1,
-    unsigned int nTri,
+    size_t nTri,
     const double* aVal)
 {
   const unsigned int np = nXY;
@@ -291,9 +291,9 @@ void MergeLinSys_SolidLinear_NewmarkBeta_MeshTri2D(
     const double gamma_newmark,
     const double beta_newmark,
     const double* aXY1,
-    unsigned int nXY,
+    size_t nXY,
     const unsigned int* aTri1,
-    unsigned int nTri,
+    size_t nTri,
     const double* aVal,
     const double* aVelo,
     const double* aAcc)
